more modifications to support split in multiple classes/commands
